Fix Eslint errors

+++ src/app.js
@@ -1,7 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import http from 'http';
-import morgan from 'morgan';
 import Database from './persistence/db';
 import api from './routes/api';
 import config from './constants/config';
@@ -19,38 +18,31 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 const db = new Database();
 
-db.connect()
-  .then(() => {
-    if (config.database.reseed) db.drop();
+if (process.env.NODE_ENV === 'TEST') {
+  db.connect()
+    .then(() => {
+      if (config.database.reseed) db.drop();
 
-    return db.getAdminsCount().
-      then((adminsCount)=>{
-        if (adminsCount === 0) 
-          return db.insertOneAdmin(config.masterAdmin);
-        return Promise.resolve()
-      })
-      .then(()=>{
-        // DUMMY DATA FOR TESTING
-        // return Promise.all([
-        //    db.insertClients(clients),
-        //    db.insertBusinesses(businesses)
-        // ])
-        return Promise.resolve()
-      })
+      return db.getAdminsCount();
+    })
+    .then((adminsCount) => {
+      if (adminsCount === 0) return db.insertOneAdmin(config.masterAdmin);
+      return Promise.resolve();
+    })
+    .then(() => Promise.all([
+      db.insertClients(clients),
+      db.insertBusinesses(businesses),
+    ]))
+    .then(() => {
+      app.use('/', express.static('public'));
+      app.use('/api/', api({ db }));
 
-
-   
-  })
-  .then(() => {
-    app.use('/', express.static('public'));
-    app.use('/api/', api({ db }));
-
-    app.server.listen(process.env.PORT || config.server.port);
-    console.log(`Server listening on port ${app.server.address().port}...`);
-  })
-  .catch((err) => {
-    console.error(err.message);
-  });
+      app.server.listen(process.env.PORT || config.server.port);
+      console.log(`Server listening on port ${app.server.address().port}...`);
+    })
+    .catch((err) => {
+      console.error(err.message);
+    });
+}
 
 app.use('/api/', api({ db }));
-export default app;
